+++ Package.swift
@@ -2,8 +2,8 @@
 
 import PackageDescription
 
-let tag = "v0.1.0-rc47"
-let checksum = "1d1fc858bd1712ee0cd5965e1839bce502a1209dbd42a09cdb9495f0f8256f6f"
+let tag = "v0.1.0-rc48"
+let checksum = "2c46ecadae1c648731dff64770e6334a127469d031bbf499135db972d9982c81"
 let url = "https://github.com/pubky/paykit-rs/releases/download/\(tag)/Paykit.xcframework.zip"
 
 let package = Package(